Move discoveredLdLinkerSpecInfo_Windows to platform tests.
* Move discoveredLdLinkerSpecInfo_Windows to the windows platform
tests.
* Use the helper functions to make linker tests conditional on the
  presence of specific host SDK components.
* Update target architectures to match windows SDK.
